Implement configurable data migration using SDK search API (#2927)
* Implement configurable data migration using SDK search API

Example use case: Fix an issue with extraction for instance this can
be used to fix incorreclty extract reference. Configuration accepts
data queries and resource type to search then applies the update values
on the resources. This function should be blocking and should run immediately
after login.

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Fix commit

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Use DefaultRepository instead of FHIREngine

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Save latest migration version

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Migrate to use PreferenceDataStore

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Fix failing test.

Ensure that the exact JSON path is used when declaring jsonPathExpression
otherwise the value for the key will just be overridden.

Example: Assuming we want to fix an issue with Task.basedOn reference
if we use this '$.basedOn[0]' or 'Task.basedOn[0]' as jsonPathExpression,
it will update the Task.basedOn reference with a string ie.g it will override
the JSON "basedOn": [ {"reference" : "CarePlan/some-uuid"} ] with
"basedOn": [ "CarePlan/some-uuid" ] since we are changing the first item in
the list. This will fail validation. Always provide the exact path for the
item to be replaced to avoid this error.

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Refactor loader dialog

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Display a blocking progress dialog when migrating data

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Run spotlessApply

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Fix LoginScreenText

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

---------

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>