Fix data migration issues (#2969)
* Fix update data migration version

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Purge resources to be updated from LocalChangeEntity table

Any operation applied to a resource is captured in the LocalChangeEntity
table. The LocalChangeEntity accepts these action types, INSERT, UPDATE
and DELETE. The types correspond to the following HttpVerbs PUT, PATCH
and DELETE. PUT ensures the server performs upsert operation on the
change made on the resources captured in the local cahnge entity table.

The purge operation is configurable; defaults to false.

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Apply filter on selected data migration resource

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Reset configs cache on load

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Update logic for change managing entity for a Group resource

Use the configured relationship code to retrieve the RelatedPerson
resource representing the managing entity from the extracted
resources.

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Fix condition

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Fix update managing entity

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Fix change managing entity for Group

Only append organization info to the relevant Resource property
if and only if the property is null. E.g. if Group.managingEntity
is already set during extraction, do not override it.

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Set Practitioner if not exist

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

* Fix failing tests

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>

---------

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>
Co-authored-by: Francis Odhiambo <4540684+f-odhiambo@users.noreply.github.com>